backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
openai==1.51.2
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
PyMuPDF==1.24.10
pydantic==2.9.2
email-validator==2.2.0
supabase==2.4.0
postgrest==0.11.1
realtime==1.0.0
gotrue==2.4.1





